+++ src/document/mod.rs
@@ -44,6 +44,7 @@ pub mod nodes_run;
 pub mod project;
 pub mod proposal;
 pub mod source;
+mod source_format;
 pub mod var_model;
 pub mod variable;